drivers: fix overshift and overflow errors on 8Bit MCUs
    - bmp180: fix overshift error
    - hih6130: fix overflow in sleep interval
    - pn532: fix overshift error
    - bh1750fvi: fix overshift error
    - mpu9050: fix overflow error

+++ drivers/mpu9150/mpu9150.c
@@ -375,8 +375,8 @@ int mpu9150_read_temperature(const mpu9150_t *dev, int32_t *output)
     /* Release the bus */
     i2c_release(dev->i2c_dev);
 
-    temp = (data[0] << 8) | data[1];
-    *output = ((((int32_t)temp) * 1000) / 340) + (35*1000);
+    temp = ((uint16_t)data[0] << 8) | data[1];
+    *output = (((int32_t)temp * 1000LU) / 340) + (35 * 1000LU);
 
     return 0;
 }
